Moved landing page components to app directory
