+++ app/components/Hero/index.tsx
