+++ app/page.tsx
@@ -1,8 +1,8 @@
 import { NextPage } from "next";
-import Hero from "../src/components/landing/Hero";
+import Hero from "./components/Hero";
 import Image from "next/image";
-import Features from "../src/components/landing/Features";
-import Faq from "../src/components/landing/Faq";
+import Features from "./components/Features";
+import Faq from "./components/Faq";
 import telemetryDemo from "../public/images/readme/telemetry-demo.png";
 
 const Home: NextPage = () => {
