+++ src/pages/season-hub/[event].tsx
@@ -4,7 +4,7 @@ import { GrandPrix, Season } from "../../model/Season";
 import { slugify } from "../../utils/slugify";
 import Navbar from "../../components/landing/Navbar/Navbar";
 import GrandPrixInfoHeader from "../../components/season-hub/GrandPrixInfoHeader/GrandPrixInfoHeader";
-import Footer from "../../components/landing/Footer/Footer";
+import Footer from "../../../app/components/Footer/Footer";
 import Table from "../../components/standings/Table/Table";
 import { formatTime } from "../../utils/formatTime";
 import { getCountryFlag } from "../archive/[season]";
